Factor out metadata to common center class
