+++ common/src/python/centers/nacc_group.py
@@ -3,11 +3,11 @@
 import logging
 from typing import Dict, Optional
 
+from centers.center_adaptor import CenterAdaptor
 from centers.center_group import CenterGroup
 from flywheel.models.group import Group
 from flywheel.models.user import User
-from flywheel_adaptor.flywheel_proxy import (FlywheelProxy, GroupAdaptor,
-                                             ProjectAdaptor)
+from flywheel_adaptor.flywheel_proxy import FlywheelProxy
 from pydantic import BaseModel, ValidationError
 
 log = logging.getLogger(__name__)
@@ -52,11 +52,10 @@ class CenterMapInfo(BaseModel):
         return self.centers.get(adcid, None)
 
 
-class NACCGroup(GroupAdaptor):
+class NACCGroup(CenterAdaptor):
     """Manages group for NACC."""
 
     def __init__(self, *, group: Group, proxy: FlywheelProxy) -> None:
-        self.__metadata: Optional[ProjectAdaptor] = None
         super().__init__(group=group, proxy=proxy)
 
     @classmethod
@@ -78,19 +77,6 @@ class NACCGroup(GroupAdaptor):
         metadata_project.add_admin_users(admin_group.get_user_access())
 
         return admin_group
-
-    def get_metadata(self) -> ProjectAdaptor:
-        """Returns the metadata project.
-
-        Returns:
-          The metadata object
-        """
-        if not self.__metadata:
-            self.__metadata = self.get_project('metadata')
-            assert self.__metadata, ("Expecting metadata project. "
-                                     "Check user has permissions.")
-
-        return self.__metadata
 
     def add_center(self, center_group: CenterGroup) -> None:
         """Adds the metadata for the center.
